-moved spell order detail to /client/components/user-orders.js
