+++ client/components/index.js
@@ -5,6 +5,7 @@
  */
 export {default as Navbar} from './navbar'
 export {default as UserHome} from './user-home'
+export {default as UserOrders} from './user-orders'
 export {default as AllSpells} from './allspells'
 export {default as AddSpell} from './addSpell'
 export {default as EditSpell} from './editspell'
