+++ client/components/navbar.js
@@ -12,6 +12,7 @@ const Navbar = ({handleClick, isLoggedIn}) => (
         <div>
           {/* The navbar will show these links after you log in */}
           <Link to="/home">Home</Link>
+          <Link to="/orders">Your Orders</Link>
           <a href="#" onClick={handleClick}>
             Logout
           </a>
